# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List practice cases
      parameters:
        - name: areaOfLaw
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AreaOfLaw"
        - name: difficulty
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Difficulty"
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
  /cases/generate:
    post:
      operationId: generateCase
      tags: [cases]
      summary: Generate a new AI-authored case scenario under Pakistani law
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseGenerateInput"
      responses:
        "201":
          description: Generated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AreaOfLaw:
      type: string
      enum:
        - Criminal
        - Civil
        - Constitutional
        - Family
        - Contract
        - Property
        - Corporate
        - Tort
    Difficulty:
      type: string
      enum:
        - Beginner
        - Intermediate
        - Advanced
    CaseSource:
      type: string
      enum:
        - library
        - generated
...
    Witness:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        statement:
          type: string
      required:
        - name
        - role
        - statement
    StatuteCitation:
      type: object
      description: A provision confirmed to exist in the statute corpus.
      properties:
        citation:
          type: string
          description: Canonical citation, e.g. "QSO 1984 Art. 71"
        statuteCode:
          type: string
        heading:
          type: string
        verified:
          type: boolean
          description: |
            True when the corpus text has been diffed against the official
            source. False means the provision exists but its wording has not
            yet been authenticated.
      required:
        - citation
        - statuteCode
        - heading
        - verified
...
    Case:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        areaOfLaw:
          $ref: "#/components/schemas/AreaOfLaw"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        summary:
          type: string
        applicableLaws:
          type: string
        petitionerName:
          type: string
        petitionerRole:
          type: string
        respondentName:
          type: string
        respondentRole:
          type: string
        witnesses:
          type: array
          items:
            $ref: "#/components/schemas/Witness"
        citations:
          type: array
          description: |
            Provisions this case is grounded in, resolved against the statute
            corpus. Empty for library cases predating statute grounding.
          items:
            $ref: "#/components/schemas/StatuteCitation"
        brief:
          description: |
            The pleading behind the case: numbered facts, lettered grounds, and
            the itemised prayer. Null for library cases and for cases generated
            before the brief existed.
          oneOf:
            - $ref: "#/components/schemas/CaseBriefDetail"
            - type: "null"
        source:
          $ref: "#/components/schemas/CaseSource"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - areaOfLaw
        - difficulty
        - summary
        - applicableLaws
        - petitionerName
        - petitionerRole
        - respondentName
        - respondentRole
        - witnesses
        - citations
        - brief
        - source
        - createdAt
    BriefItem:
      type: object
      description: One numbered fact, lettered ground, or itemised prayer.
      properties:
        label:
          type: string
          description: |
            Assigned server-side ("1", "A", "a") so labels stay contiguous when
            an item is dropped by the citation audit.
        text:
          type: string
      required:
        - label
        - text
    CaseParty:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        description:
          type: string
      required:
        - name
        - role
        - description
    CaseBriefDetail:
      type: object
      description: |
        The case as a filing would plead it. `summary` remains the short
        orientation paragraph; this is what the student argues from — grounds are
        the argument, prayer is the relief sought. Every ground's citations are
        audited against the statute corpus before the case is stored.
      properties:
        court:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        jurisdictionInvoked:
          type: ["string", "null"]
        petitioners:
          type: array
          items:
            $ref: "#/components/schemas/CaseParty"
        respondents:
          type: array
          items:
            $ref: "#/components/schemas/CaseParty"
        facts:
          type: array
          items:
            $ref: "#/components/schemas/BriefItem"
        grounds:
          type: array
          items:
            $ref: "#/components/schemas/BriefItem"
        prayer:
          type: array
          items:
            $ref: "#/components/schemas/BriefItem"
      required:
        - court
        - caseNumber
        - jurisdictionInvoked
        - petitioners
        - respondents
        - facts
        - grounds
        - prayer
    CaseGenerateInput:
      type: object
      properties:
        areaOfLaw:
          $ref: "#/components/schemas/AreaOfLaw"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
      required:
        - areaOfLaw
        - difficulty
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error