# lib/api-zod/src/generated/types/briefItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * One numbered fact, lettered ground, or itemised prayer.
 */
export interface BriefItem {
  /**
     * Assigned server-side ("1", "A", "a") so labels stay contiguous when
     * an item is dropped by the citation audit.
     */
  label: string;
  text: string;
}