# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AreaOfLaw } from './areaOfLaw';
import type { CaseBriefDetail } from './caseBriefDetail';
import type { CaseSource } from './caseSource';
import type { Difficulty } from './difficulty';
import type { StatuteCitation } from './statuteCitation';
import type { Witness } from './witness';

export interface Case {
  id: number;
  title: string;
  areaOfLaw: AreaOfLaw;
  difficulty: Difficulty;
  summary: string;
  applicableLaws: string;
  petitionerName: string;
  petitionerRole: string;
  respondentName: string;
  respondentRole: string;
  witnesses: Witness[];
  /**
     * Provisions this case is grounded in, resolved against the statute
     * corpus. Empty for library cases predating statute grounding.
     */
  citations: StatuteCitation[];
  /**
     * The pleading behind the case: numbered facts, lettered grounds, and
     * the itemised prayer. Null for library cases and for cases generated
     * before the brief existed.
     */
  brief: CaseBriefDetail | null;
  source: CaseSource;
  createdAt: Date;
}