# lib/api-zod/src/generated/types/caseBriefDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BriefItem } from './briefItem';
import type { CaseParty } from './caseParty';

/**
 * The case as a filing would plead it. `summary` remains the short
 * orientation paragraph; this is what the student argues from — grounds are
 * the argument, prayer is the relief sought. Every ground's citations are
 * audited against the statute corpus before the case is stored.
 */
export interface CaseBriefDetail {
  /** @nullable */
  court: string | null;
  /** @nullable */
  caseNumber: string | null;
  /** @nullable */
  jurisdictionInvoked: string | null;
  petitioners: CaseParty[];
  respondents: CaseParty[];
  facts: BriefItem[];
  grounds: BriefItem[];
  prayer: BriefItem[];
}